refactor: extract item selector type-to-search lifecycle
